backend/extraction_service/requirements.txt
# Web framework and related
fastapi==0.111.1
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==1.10.13

# PDF and document processing
PyPDF4==1.27.0
python-docx==1.1.2
pdfplumber==0.11.3
nltk==3.9.1

# File operations
aiofiles==23.2.1

# Path handling
pathlib==1.0.1

# Utilities
typing-extensions==4.12.0